Amended the package controller 'on_start' public function
Previously this function was commented out as there's currently an error where this function requires commenting out to uninstall the package.  Changes made to the app name and path within this function to match the perceived requirements and current testing.  Currently this results in an error (Fatal error: Class 'Concrete\Package\XyzSpotifyAuthType\Core' not found) which is yet to be resolved.